New main definitions file
