+++ libs/aggiemap/ngx/common/src/lib/layer-sources/layer-sources.ts
@@ -1,18 +1,11 @@
 import { LayerSource } from '@tamu-gisc/common/types';
 
-import { IComposedIDefinitions } from '../definitions';
 import { IComposedConnections } from '../connections';
+import { IComposedIDefinitions } from '../definitions';
 import { IFactoryExcludeOptions } from '../utils/definitionFactory';
+import { MainMapLayerSources } from '../definitions/main.definitions';
 
-import esri = __esri;
-
-export const commonLayerProps = {
-  outFields: ['*'],
-  minScale: 100000,
-  maxScale: 0,
-  elevationInfo: { mode: 'relative-to-ground', offset: 1 } as esri.FeatureLayerElevationInfo,
-  popupEnabled: false
-};
+export { commonLayerProps } from '../definitions/main.definitions';
 
 // Persistent layer definitions that will be processed by a factory and added to the map.
 export function LayerSources(
@@ -20,452 +13,7 @@ export function LayerSources(
   definitions: IComposedIDefinitions,
   options?: IFactoryExcludeOptions<IComposedIDefinitions>
 ): Array<LayerSource> {
-  const all: Array<LayerSource> = [
-    {
-      type: 'feature',
-      id: definitions.BUILDINGS.layerId,
-      title: definitions.BUILDINGS.name,
-      url: definitions.BUILDINGS.url,
-      popupComponent: definitions.BUILDINGS.popupComponent,
-      listMode: 'hide',
-      visible: true,
-      essential: true,
-      layerIndex: 1,
-      native: {
-        ...commonLayerProps,
-        legendEnabled: false,
-        renderer: {
-          type: 'simple',
-          symbol: {
-            type: 'simple-fill',
-            style: 'solid',
-            color: [0, 0, 0, 0.01],
-            outline: {
-              width: '0'
-            }
-          }
-        }
-      }
-    },
-    {
-      type: 'feature',
-      id: definitions.CONSTRUCTION.layerId,
-      title: definitions.CONSTRUCTION.name,
-      url: definitions.CONSTRUCTION.url,
-      popupComponent: definitions.CONSTRUCTION.popupComponent,
-      listMode: 'show',
-      visible: false,
-      essential: false,
-      layerIndex: 2,
-      native: {
-        ...commonLayerProps,
-        definitionExpression: `EndDate > CAST('${new Date().toISOString()}' AS DATE ) AND Status = 'Active'`
-      }
-    },
-    {
-      type: 'feature',
-      id: definitions.POINTS_OF_INTEREST.layerId,
-      title: definitions.POINTS_OF_INTEREST.name,
-      url: definitions.POINTS_OF_INTEREST.url,
-      popupComponent: definitions.POINTS_OF_INTEREST.popupComponent,
-      listMode: 'show',
-      visible: false,
-      native: {
-        ...commonLayerProps,
-        renderer: {
-          type: 'simple',
-          symbol: {
-            type: 'picture-marker',
-            url: '/assets/images/markers/statue-icon.png',
-            width: '20px',
-            height: '30.2px'
-          }
-        }
-      }
-    },
-    {
-      type: 'feature',
-      id: definitions.BONFIRE.layerId,
-      title: definitions.BONFIRE.name,
-      url: definitions.BONFIRE.url,
-      popupComponent: definitions.BONFIRE.popupComponent,
-      listMode: 'hide',
-      visible: true,
-      native: {
-        ...commonLayerProps,
-        labelingInfo: [],
-        renderer: {
-          type: 'simple',
-          symbol: {
-            type: 'simple-fill',
-            color: [0, 0, 0, 0.0],
-            outline: {
-              width: 0
-            }
-          }
-        }
-      }
-    },
-    {
-      type: 'feature',
-      id: definitions.LACTATION_ROOMS.layerId,
-      title: definitions.LACTATION_ROOMS.name,
-      url: definitions.LACTATION_ROOMS.url,
-      popupComponent: definitions.LACTATION_ROOMS.popupComponent,
-      listMode: 'show',
-      visible: false,
-      native: {
-        ...commonLayerProps
-      }
-    },
-    {
-      type: 'feature',
-      id: definitions.SURFACE_LOTS.layerId,
-      title: definitions.SURFACE_LOTS.name,
-      url: definitions.SURFACE_LOTS.url,
-      popupComponent: definitions.SURFACE_LOTS.popupComponent,
-      listMode: 'hide',
-      visible: true,
-      layerIndex: 1,
-      native: {
-        ...commonLayerProps,
-        legendEnabled: false,
-        opacity: 0.0,
-        labelingInfo: [
-          {
-            symbol: {
-              type: 'text',
-              color: [0, 0, 0, 0]
-            }
-          }
-        ]
-      }
-    },
-    {
-      type: 'feature',
-      id: definitions.VISITOR_PARKING.layerId,
-      title: definitions.VISITOR_PARKING.name,
-      url: definitions.VISITOR_PARKING.url,
-      popupComponent: definitions.VISITOR_PARKING.popupComponent,
-      listMode: 'show',
-      visible: false,
-      native: {
-        ...commonLayerProps
-      }
-    },
-    {
-      type: 'feature',
-      id: definitions.ACESSIBLE_ENTRANCES.layerId,
-      title: definitions.ACESSIBLE_ENTRANCES.name,
-      url: definitions.ACESSIBLE_ENTRANCES.url,
-      popupComponent: definitions.ACESSIBLE_ENTRANCES.popupComponent,
-      listMode: 'show',
-      visible: false,
-      native: {
-        ...commonLayerProps
-      }
-    },
-    {
-      type: 'feature',
-      id: definitions.EMERGENCY_PHONES.layerId,
-      title: definitions.EMERGENCY_PHONES.name,
-      url: definitions.EMERGENCY_PHONES.url,
-      listMode: 'show',
-      visible: false,
-      native: {
-        ...commonLayerProps
-      }
-    },
-    {
-      type: 'graphics',
-      id: 'selection-layer',
-      title: 'Selected Buildings',
-      listMode: 'hide',
-      visible: true,
-      essential: true,
-      popupComponent: definitions.BUILDINGS.popupComponent,
-      native: {
-        ...commonLayerProps
-      }
-    },
-    {
-      type: 'graphics',
-      id: 'bus-route-layer',
-      title: 'Bus Routes',
-      listMode: 'hide',
-      visible: true,
-      native: {
-        ...commonLayerProps
-      }
-    },
-    {
-      type: 'geojson',
-      id: definitions.BIKE_LOCATIONS.layerId,
-      title: definitions.BIKE_LOCATIONS.name,
-      url: definitions.BIKE_LOCATIONS.url,
-      listMode: 'show',
-      visible: false,
-      native: {
-        ...commonLayerProps,
-        renderer: {
-          type: 'simple',
-          symbol: {
-            type: 'simple-marker',
-            style: 'circle',
-            size: 8,
-            color: '#03C4A6'
-          }
-        }
-      }
-    },
-    {
-      type: 'group',
-      id: 'sustainable-transportation-group-layer',
-      title: 'Sustainable Transportation',
-      listMode: 'show',
-      visible: false,
-      sources: [
-        {
-          type: 'feature',
-          id: definitions.BIKE_DISMOUNT_ZONES.layerId,
-          title: definitions.BIKE_DISMOUNT_ZONES.name,
-          url: definitions.BIKE_DISMOUNT_ZONES.url,
-          listMode: 'show',
-          visible: true,
-          popupComponent: definitions.BIKE_DISMOUNT_ZONES.popupComponent,
-          popupData: {
-            name: '{attributes.Name}',
-            description: '<strong>Notes</strong>: {attributes.Bike_Notes}'
-          },
-          native: {
-            ...commonLayerProps
-          }
-        },
-        {
-          type: 'feature',
-          id: definitions.CITY_BIKE_LANES_ROUTES.layerId,
-          title: definitions.CITY_BIKE_LANES_ROUTES.name,
-          url: definitions.CITY_BIKE_LANES_ROUTES.url,
-          listMode: 'show',
-          visible: true,
-          popupComponent: definitions.CITY_BIKE_LANES_ROUTES.popupComponent,
-          popupData: {
-            name: '{attributes.Type}',
-            description: '<strong>Location</strong>: {attributes.Loc}'
-          },
-          native: {
-            ...commonLayerProps
-          }
-        },
-        {
-          type: 'feature',
-          id: definitions.CAMPUS_BIKE_LANES.layerId,
-          title: definitions.CAMPUS_BIKE_LANES.name,
-          url: definitions.CAMPUS_BIKE_LANES.url,
-          listMode: 'show',
-          visible: true,
-          popupComponent: definitions.CAMPUS_BIKE_LANES.popupComponent,
-          popupData: {
-            name: '{attributes.Use_}',
-            description: '<strong>Location</strong>: {attributes.Location}'
-          },
-          native: {
-            ...commonLayerProps
-          }
-        },
-        {
-          type: 'feature',
-          id: definitions.BIKE_FIX_STATIONS.layerId,
-          title: definitions.BIKE_FIX_STATIONS.name,
-          url: definitions.BIKE_FIX_STATIONS.url,
-          listMode: 'show',
-          visible: true,
-          popupComponent: definitions.BIKE_FIX_STATIONS.popupComponent,
-          popupData: {
-            name: '{attributes.Bike_Sta_Name}',
-            description: '<strong>Amenities</strong>: {attributes.Bike_Amenities}'
-          },
-          native: {
-            ...commonLayerProps
-          }
-        },
-        {
-          type: 'feature',
-          id: definitions.BIKE_RACKS_MAP.layerId,
-          title: definitions.BIKE_RACKS_MAP.name,
-          url: definitions.BIKE_RACKS_MAP.url,
-          listMode: 'show',
-          visible: true,
-          popupComponent: definitions.BIKE_RACKS_MAP.popupComponent,
-          popupData: {
-            name: '{attributes.Type}',
-            description:
-              '<strong>Total Capacity</strong>: {attributes.Total_Capacity}\n' +
-              '<strong>Notes</strong>: {attributes.BR_Notes}'
-          },
-          native: {
-            ...commonLayerProps
-          }
-        },
-        {
-          type: 'feature',
-          id: definitions.EV_CHARGE_STATIONS.layerId,
-          title: definitions.EV_CHARGE_STATIONS.name,
-          url: definitions.EV_CHARGE_STATIONS.url,
-          listMode: 'show',
-          visible: true,
-          popupComponent: definitions.EV_CHARGE_STATIONS.popupComponent,
-          popupData: {
-            name: '{attributes.EV_ID}',
-            description:
-              '<strong>Network</strong>: {attributes.EV_Network}\n' +
-              '<strong>Charging Level</strong>: {attributes.Ch_Level}\n' +
-              '<strong>Garage Level</strong>: {attributes.Garage_Lvl}\n' +
-              '<strong>Notes</strong>: {attributes.EVCS_Notes}'
-          },
-          native: {
-            ...commonLayerProps
-          }
-        }
-      ],
-      native: {
-        listMode: 'hide-children'
-      }
-    },
-    {
-      type: 'geojson',
-      id: definitions.DINING_LOCATIONS.layerId,
-      title: definitions.DINING_LOCATIONS.name,
-      url: definitions.DINING_LOCATIONS.url,
-      listMode: 'show',
-      visible: false,
-      popupComponent: definitions.DINING_LOCATIONS.popupComponent,
-      native: {
-        ...commonLayerProps,
-        renderer: {
-          type: 'unique-value',
-          field: 'label',
-          field2: 'type',
-          fieldDelimiter: ',',
-          uniqueValueInfos: [
-            {
-              value: 'open,food-truck',
-              label: 'Food Truck - Open',
-              symbol: {
-                type: 'picture-marker',
-                url: '/assets/images/icons/FoodTruck_open.png',
-                width: '24px',
-                height: '32px'
-              }
-            },
-            {
-              value: 'closed,food-truck',
-              label: 'Food Truck - Closed',
-              symbol: {
-                type: 'picture-marker',
-                url: '/assets/images/icons/FoodTruck_closed.png',
-                width: '24px',
-                height: '32px'
-              }
-            },
-            {
-              value: 'open,fixed',
-              label: 'Dining - Open',
-              symbol: {
-                type: 'picture-marker',
-                url: '/assets/images/icons/Dining_open.png',
-                width: '24px',
-                height: '32px'
-              }
-            },
-            {
-              value: 'closed,fixed',
-              label: 'Dining - Closed',
-              symbol: {
-                type: 'picture-marker',
-                url: '/assets/images/icons/Dining_closed.png',
-                width: '24px',
-                height: '32px'
-              }
-            }
-          ]
-        } as unknown as esri.UniqueValueRenderer
-      }
-    },
-    {
-      type: 'feature',
-      id: definitions.AGGIEPRINT_LOCATIONS.layerId,
-      title: definitions.AGGIEPRINT_LOCATIONS.name,
-      url: definitions.AGGIEPRINT_LOCATIONS.url,
-      listMode: 'show',
-      visible: false,
-      popupComponent: definitions.AGGIEPRINT_LOCATIONS.popupComponent,
-      popupData: {
-        description: `<strong>Access</strong>: {attributes.Access}\n<strong>Building</strong>: {attributes.BuildingName} ({attributes.BuildingNumber})\n<strong>Printer Type</strong>: {attributes.PrinterType}\n<strong>Details</strong>: {attributes.PrinterDetails}`
-      },
-      native: {
-        ...commonLayerProps,
-        renderer: {
-          type: 'unique-value',
-          valueExpression: `When($feature.Access == 'Campus Member Accessible', 'all',  'restricted')`,
-          uniqueValueInfos: [
-            {
-              value: 'all',
-              label: 'Campus Member Accessible',
-              symbol: {
-                type: 'picture-marker',
-                url: '/assets/images/icons/services/printer-all-access.png',
-                width: '24px',
-                height: '32px'
-              }
-            },
-            {
-              value: 'restricted',
-              label: 'Restricted Access Printers',
-              symbol: {
-                type: 'picture-marker',
-                url: '/assets/images/icons/services/printer-restricted.png',
-                width: '24px',
-                height: '32px'
-              }
-            }
-          ]
-        } as unknown as esri.UniqueValueRenderer
-      }
-    },
-    {
-      type: 'feature',
-      id: definitions.SINGLE_OCCUPANCY_RESTROOMS.layerId,
-      title: definitions.SINGLE_OCCUPANCY_RESTROOMS.name,
-      url: definitions.SINGLE_OCCUPANCY_RESTROOMS.url,
-      popupComponent: definitions.SINGLE_OCCUPANCY_RESTROOMS.popupComponent,
-      listMode: 'show',
-      visible: false,
-      popupData: {
-        description:
-          '<strong>Building</strong>: {attributes.Name}\n' + '<strong>Restroom Location(s)</strong>: {attributes.Notes}'
-      },
-      native: {
-        ...commonLayerProps,
-        definitionExpression: "showOnAggieMap = 'Y'"
-      }
-    }
-  ];
-
-  return all.filter((source) => {
-    // Filter out any sources that are in the exclude list
-    if (options && options.exclude && options.exclude.length > 0) {
-      const keyId = options.exclude.some((key) => {
-        return definitions[key].layerId === source.id;
-      });
-
-      if (keyId) {
-        return false;
-      }
-    }
-
-    return true;
-  });
+  return MainMapLayerSources(connections, definitions, options);
 }
 
 export const ThreeDLayers: Array<LayerSource> = [
@@ -506,7 +54,7 @@ export const ThreeDLayers: Array<LayerSource> = [
             }
           ]
         }
-      } as any
-    }
+      }
+    } as unknown as Record<string, unknown>
   }
 ];
